SAM-96
Issue 3: Family >details>Members Details : Birth Date format should be appear like web.

Issue 4: Family >details>Members Details Hight, weight ,income, when added with decimal value in web in mobile app still showing normal value.

Issue 5:   Family >details>Members Details : Education and Occupation related information not appear members details.

Issue 6:  Family >details>Members Details : Add space between Home and Address section.

Issue 5:  Family >details>Members Details : remove seconds from Birth time.

Issue 6:   Family >details>Members Details : when social medias link should be appear in 1 line.

+++ lib/features/member/domain/entities/member.dart
@@ -159,7 +159,7 @@ class Member {
       educationName: getString('educationName', 'educationalQualification', 'education'),
       jobPositionName: getString('jobPositionName'),
       otherJobPosition: getString('otherJobPosition'),
-      monthlyIncome: (json['monthlyIncome'] as num?)?.toInt(),
+      monthlyIncome: (json['monthlyIncome'] as num?)?.toDouble(),
       isOwnLand: json['isOwnLand'] as bool?,
       isOwnHouse: json['isOwnHouse'] as bool?,
       hasTwoWheeler: json['hasTwoWheeler'] as bool?,
@@ -251,7 +251,7 @@ class Member {
   final String? educationName;
   final String? jobPositionName;
   final String? otherJobPosition;
-  final int? monthlyIncome;
+  final double? monthlyIncome;
   final bool? isOwnLand;
   final bool? isOwnHouse;
   final bool? hasTwoWheeler;
